+++ workspaces/x2a/plugins/x2a/src/components/tools/index.ts
@@ -15,3 +15,4 @@
  */
 export * from './buildArtifactUrl';
 export * from './humanizeArtifactType';
+export * from './humanizeDate';